Fix async invitation acceptance bug
When you call `Core::AcceptInvitation()`, you register a callback that gets
invoked asynchronously on the "origin thread" (typically UI thread). The
callback accepts a MageHandle, which is the handle associated with the finished
invitation. It represents the primordial connection to the inviter process.
Before the IO thread posts a task to invoke the callback on the origin thread,
`Core::OnReceivedInvitation` generates a new MageHandle associated with the
newly-created `Endpoint`. It passes this handle to the UI thread in the task
that gets invoked there. Previously, we "passed" the value of the handle by
reference, even though we were inside a function where the handle value would
likely be out of scope by the time the task ran on the UI thread. The value of
the local handle was corrupt by the time it was being read on the UI thread,
defaulting to 0, and causing a CHECK() failure when trying to actually use the
handle.

This commit captures the local handle by value in the task sent to the UI
thread, thus fixing the issue.