MageHandle passing: Process `EndpointInfo`s on IO thread
This commit makes `Endpoint` process the various `EndpointInfo`s inside a
`Message` on the IO thread, before we ever dispatch the message to the receiver
delegate. This fixes a race condition that we previously saw, where if a parent
process sends two messages
  1.) [InterfaceA]: Message(InterfaceB handle)
  2.) [InterfaceB]: Message(...)

The receiving process would:
  1.) Process the first message on the IO thread
  2.) Post a task to dispatch the message to the receiving delegate. This task
      unpack the parameters of the thread, which includes creating any concrete
      `Endpoint`s that are described by `EndpointInfo`s in the message buffer
  3.) Immediately process the second message on the IO thread, that is bound for
      an endpoint that will be asynchronously created in the task above.

By creating a standardized place in `MessageHeader` to reference all
`EndpointInfo`s from, we allow non-generated code on the IO thread to unpack all
`EndpointInfo`s in a user message, and create the concrete `Endpoint`s for them
before the next message will be read (since the next message might target one of
the newly-created `Endpoint`s).

There are still at least three bugs with the implementation:
  1.) If `Endpoint::AcceptMessage` queues a message that has handles, we will
      create the `Endpoint`s for those handles, queue the message, and then
      later re-enter `AcceptMessage()` and create the handles again, which is a
      bug. We should have a test for this.
  2.) The example mage_parent.cc relies on a thread sleep before sending a
      message over a handle which was just sent over an existing handle. This
      should not be necessary. We should removing the sleep and fix the bug.
  3.) Synchronization issue: Since `Endpoint::state` and
      `Endpoint::incoming_messages_` can be accessed unsafely from multiple
      threads, it is possible to queue a message to `incoming_messages_` at the
      same time a receiver delegate gets bound. The result is the freshly-bound
      delegate sees no recently-queued messages to dispatch, and then
      `incoming_messages_` ends up never delivering its messages.