Enable "recursive" message sending for queued endpoints
This commit allows us to remove the delay in `examples/mage/mage_parent`, so
that we can start using the message pipe associated with the invitation
immediately after we send the invitation, without waiting for the invitation to
be accepted. We could already do that for normal messages, but for messages that
carried EndpointDescriptors this didn't work. This commit makes this work even
for messages carrying EndpointDescriptors, by implementing a BFS search for the
transitive closure of all messages that should be sent when some "parent"
message gets sent to another process. This solves problem (2) described in
commit `a653fd7`.

There are still two remaining issues with message queueing:
  1.) Messages sent on "child" pipes of the invitation pipe *after* the
      invitation is accepted will likely never be delivered. We should have a
      test for this and fix it.
  2.) Generalized message queueing for messages with EndpointDescriptors is
      broken. It is maybe even broken for messages that don't carry
      EndpointDescriptors. We should have tests for all of these cases.