Enable multiple MageHandles in a message, sending messages back to parent
This commit enables the ability to fully utilize multiple MageHandles in a
message, as demonstrated by the principal mage example where we send two
messages to another process. One handle gets messages queued on it for a
receiver bound in that remote process, and the other is used for a remote to
message back the parent process that it is finished.