Fix bug with queueing EndpointDescriptor-bearing message
This commit fixes a bug where `Endpoint::AcceptMessage()` would unconditionally
recover new `Endpoint`s from the `EndpointDescriptor`s in messages. This meant
that for handle-bearing messages sent to *local* endpoints via the *local* path
in `Node`, we'd still create and register duplicate endpoints for the
descriptors, despite them already existing by definition (again, the receiving
endpoint was local). To fix this we introduce
`Endpoint::AcceptMessageOnIOThread()` which is only called in the cross-process
message receiving path. This method is the only one that does the processing of
the endpoints in messages.

It delegates to `mage::Core::RecoverMageHandleFromEndpointDescriptor()` to
create new endpoints from the `EndpointDescriptor`s. As part of this, this
commit makes `Core::RegisterLocalHandle()` check that the endpoint being added
does not previously exist. This required changes to
`Node::SendInvitationAndGetMessagePipe()`, which previously did the following:
  1.) Created its own `Endpoint`s
  2.) Called `InitializeAndEntangleEndpoints()`, which would insert the
      endpoints into `local_endpoints_`
  3.) Called `Core::Get()->RegisterLocalHandle()`, which now asserts that the
      given endpoint is not registered

This required making `InitializeAndEntangleEndpoints()` *not* insert the given
endpoint into `Node::local_endpoints_`. This work gave us the benefit of making
`Node::SendInvitationAndGetMessagePipe()` use the traditional way of creating
message pipes instead of having its own novel way.