Write queueing tests for pipes queued and forwarded by the invitation pipe
The remaining tests to write in this style are the ones for pipes not passed
over the initial invitation pipe. At this point in time, the initial invitation
pipe is special in that it is the only kind that does the right style of message
queueing and forwarding, espectially for messages that contain
`EndpointDescriptor`s (see `Node::OnReceivedAcceptInvitation`, which does
transitive BFS message queueing upon the acceptance of an invitation).

This means that any pipes created to expand the connections between two
processes *after* this point do not benefit from the same queueing. We must
generalize the logic that is currently peculiar to the initial invitation pipe
in order to get this to work. And write tests for it of course.