Generalize recursive message queueing for all cross-process messages
So-called "recursive" message queueing is when a message is queued and that
message itself carries endpoints, each of which may already be queueing their
own messages. We end up with this sort of tree of queued messages that all need
to be forwarded along with the root message whenever it is sent to a concrete
endpoint.

Before this commit, "recursive" message forwarding was implemented (with BFS)
only for messages sent on the initial invitation endpoint.

This commit generalizes recursive message forwarding to work for pipes
established after the initial child process invitation is extended, by
introducing `Node::SendMessagesAndRecursiveDependants()`. This method is called
by both `Node::OnReceivedAcceptInvitation()` and `Node::SendMessage()` (for the
local-peer-is-unbound-and-proxying case). This change is supplemented with tests
confirming the scenario.

This is a great step, but recursive message forwarding is still broken for the
same-process case. When an endpoint-baring message is received by a
`mage::Endpoint` in the same-process case, we never call
`mage::Core::RecoverMageHandleFromEndpointDescriptor()` for each endpoint and
register it with the `mage::Message` via (`mage::Message::QueueHandle()`), since
that path is designed to truly unpack and register endpoints that are foreign to
the receiving process. But since we never queue handles in the message upon
receipt, when the `ReceiverDelegate` tries to take handles from the message (via
`mage::Message::TakeNextHandle()`), we blow up. So what we need is a
light-weight version of the handle recovery steps for the same-process case.
This would unpack the endpoint descriptors, but not register the endpoints with
`Core` and `Node`, since they should already exist.