Add a core API to process-but-not-register new `EndpointDescriptor`s
This commit adds a new core method used by
`Endpoint::AcceptMessageOnDelegateThread()` that recovers existing endpoints
that back corresponding `EndpointDescriptor`s in a given message. This is
specifically used for the case where an endpoint sends a message to a peer that
happens to be local (and is not proxying). In this local-peer-endpoint case,
messages that contain `EndpointDescriptor`s must be "processed" so that we can
queue the associated `MageHandle`s (described by the descriptors) to
`Message::handles_`, without actually creating new `Endpoint`s to back the
descriptors in the message.

For example: if process A sends process B some endpoints, then upon message
receipt, process B will have to create new endpoints to back those described by
the endpoint descriptors in the given message. But if process A sends *itself*
endpoints, then when the local peer node receives the message, the endpoints
backing the descriptors should already exist in this process. This commit adds
new ways to recover existing endpoints from `EndpointDescriptor`s without
*creating* new endpoints.