Support arbitrary endpoint passing, and returning to the same process
This commit started out with at least three tests:
  1.) MageTest.ChildPassAcceptInvitationPipeBackToParent
  2.) MageTest.ChildPassSendInvitationPipeBackToParent
  3.) MageTest.ChildPassRemoteAndReceiverToParent

...which were all failing at first. This commit fixes all those tests by first
resolving the design discussion around how we'll actually handle passing
endpoints cross-process. Before, we had an issue where we could not pass an
endpoint back to a process that it had ever been before. This is because the
name of an endpoint would never change no matter what process it's in, so once
we passed an endpoint *from* a given process, it would keep the endpoint around
there in a proxying state, and would clash with the same endpoint if it ever
returned back to the process.

We fix this issue, which was causing the tests to fail in various ways, by
renaming endpoints when they are transported to another process. This is
achieved by having the sending process up-front compute a random cross-process
name for all `EndpointDescriptor`s it computes. If the endpoint actually goes
cross-process, we set the local-endpoint-representing-the-descriptor's
`proxy_target`'s `endpoint_name` to the cross-process endpoint name, so that the
sender is automatically set-up to target the newly-named endpoint that will be
created in the target node. This is a big change from a previous invariant of
having endpoints retain the same name wherever they are.

This commit also finally supports the case where an endpoint that is in the
proxying state receives a message from another node. This also involved adding a
temporary and nasty Endpoint->Core path (`Core::ForwardMessage()`) that Endpoint
uses when it receives a message in this state. This allows a proxying endpoint
to forward messages on to the proxy target, which is necessary for many reasons
including ones exercised by the test.

This commit also does one thing that should be reverted, which is that we allow
invitee nodes to choose their own primordial endpoint name, and communicate this
back to the inviter upon invitation acceptance. This means in the inviter node,
all messages that were queued in the local unbound endpoint (that is not yet put
into the proxying state) have to have their target endpoint changed. This caused
some complicated logic in `Node::OnReceivedAcceptInvitation()`. This just isn't
necessary though, instead the inviter should choose the remote endpoint's name
*at random* just like we do when we send endpoint descriptors to another
process. I'll clean this up in a follow-up commit.