Proxy new endpoints if the target endpoint is proxying
Before this commit, if a message was sent to a proxying endpoint, the message
would get appropriately forwarded to the target endpoint (that we're proxying
to). But if the message contained endpoint descriptors (which get turned into
fresh, concrete `Endpoint`s after the message was received by the proxying
endpoint), these freshly-created dependant endpoints were not set into the
proxying state. So messages bound for them would get stuck in the remote
endpoint forever.

This commit modifies `Core::ForwardMessage` to not only forward messages from a
proxying endpoint to its proxy target, but also find any endpoints that were
created by the endpoint descriptors in the message (if any) and set them to the
proxying state as well.

This commit also includes a test for this scenario.