Synchronize state in `Endpoint`
This commit fixes a common and nasty race condition in `Endpoint` whereby
`AcceptMessage()` and `RegisterDelegate()` were both being called from the IO
and UI threads respectively at the same time. When this happened,
`AcceptMessage()` would observe that `state == State::kUnboundAndQueueing`, and
it would decide to queue the message to `Endpoint::incoming_message_queue_`.
However, after it decided to queue the message but before it could actually do
so, `RegisterDelegate()` would be called and run to completion. This updated
`state == kBound`, and found no messages in `incoming_message_queue_` to post to
the newly-bound delegate. Then by the time `AcceptMessage()` actually queued the
message, it was doing so when it should have been posting the message to the
newly-bound delegate.

This resulted in the message sitting around forever and never getting sent. This
commit fixes this by adding a `base::Mutex` to `Endpoint` and synchronizing
access to the following members:
 - `state`
 - `delegate_`
 - `delegate_task_runner_`
 - `incoming_message_queue_`