Fix race condition when sending to local peer endpoint
This commit mostly fixes a race condition that was encountered when
sending a message to a local peer endpoint. When sending this message, there
were two places where we read `Endpoint::state` and acted on it, but neither
were guarded by the `Endpoint`'s internal lock. In practice, the value of the
first read should match the value of the second read (they just happened to be
separated by a function), and the correctness of the flow depends on this.
However, in the test `MageTest.ChildPassSendInvitationPipeBackToParent` we were
seeing something representative of a larger problem: on the IO thread, the
endpoint was being put into a proxying state, which was breaking things.

The fix for this is to of course guard the entire read flow and the write (which
sets the endpoint into the proxying state) behind a lock. For now, this commit
just guards the entire read flow behind a lock, and the race is mostly resolved
in practice. Subsequent commits will appropriately guard all relevant accesses
of `Endpoint` so that more complicated and subtle races cannot happen, but this
requires some architecture changes that are out-of-scope for this commit.