Remove `Core::ForwardMessage()` in favor of `Node::PrepareToForwardUserMessage()`
Before this commit, `Endpoint` was responsible for receiving messages (via
`Endpoint::AcceptMessage*()`) even when its state was
`State::kUnboundAndProxying`. In order to forward these messages, it deferred to
`Core::ForwardMessage()` which:
  1.) Prepared to forward the message, by setting each dependent Endpoint into
      the proxying state
  2.) Sent the outer message (which implicitly sent any dependent Endpoints)

Having `Endpoint` reach back out into `Core` was always a layering violation,
and as such that set up was always temporary. This commit finally removes the
temporary `Core::ForwardMessage()` in favor of
`Node::PrepareToForwardUserMessage()`, which is called directly by `Node` before
we even enter `Endpoint`. As such, `Endpoint` no longer receives any messages
while it is in the `kUnboundAndProxying` state, as that should be handled by the
layer above (`Node`).