Fix Endpoint race condition
This commit finally fixes the race condition seen earlier, that
https://github.com/domfarolino/browser/pull/32/commits/0218a356025292a24909a1b618dfd2d8c83503e0
made some progress towards fixing. This is achieved by locking `remote_endpoint`
in `Node::OnReceivedAcceptInvitation()`, so that it is impossible to transition
`remote_endpoint` into the proxying state while it is being queued to on another
thread at the same time.

This removes a big class of race conditions from mage core.