Fix race condition where dependant endpoint messages sent too early
This commit fixes what appears to be the final race condition observed in the
endpoint-managing code. The race condition was where dependant endpoints of a
particular message were set into the proxying state without being locked, which
meant that new messages could be sent on them to their proxy target *before* the
message-they-were-attached-to was sent. This meant that the target node would
receive messages bound for an endpoint that didn't exist yet, causing it to
crash.

The solution here is to lock all dependant endpoints of a particular message
before putting them into the proxy state, and only after their "host" message
was sent.