Make `Receiver::stub_` a shared_ptr for `Endpoint` to weakly reference
Before this commit, it was possible for `mage::Endpoint` to dispatch a message
to a `stub_` that was destroyed, because the `mage::Receiver` it was attached to
was destroyed as well. This would cause a crash (use-after-free), despite it
being very possible to hit this case.

After this commit, `mage::Receiver` strongly references its underlying generated
`stub_`, which passes a weak_ptr of itself to `mage::Endpoint` for keeping. When
`Endpoint` receives a message bound for its underlying `delegate_`, it posts a
task to `ReceiverDelegate` as usual, but now the task posted checks to see if
`delegate_` is alive. If it is, we dispatch the message to the generated stub
for deserialization and synchronous dispatch. Otherwise, we drop the message,
instead of naively trying to invoke a method on the destroyed stub/delegate.

This commit cleans up a few previously-nasty lifetime things from the MVP.
Specifically:
  - Receiver holds a shared_ptr to its generated stub
  - Endpoint holds a weak_ptr to its ReceiverDelegate (which is the
    aforementioned stub)
  - ReceiverDelegate now checks whether the weak_ptr to the delegate is expired,
    so it can know whether it should drop or dispatch the message