Stop calling Endpoint::SetProxying from Core::PopulateEndpointDescriptor()
This is a really nice simplification. Before this commit,
Core::PopulateEndpointDescriptor() did that, and *sometimes* set the
endpoint-being-sent into the proxying state, if it detected that the endpoing
was certainly going cross-node/process (specifically, if it detected that the
endpoint backing the preexisting connection had a remote peer; it didn't account
for the case where it had a local peer in a proxying state).

That was bad because it was perhaps the last place where we were calling
`Endpoint::SetProxying()` outside of a lock. We couldn't lock the endpoint
inside `PopulateEndpointDescriptor[...]()` either, because the required sequence
of events was:
  1.) Lock the endpoint
  2.) Call SetProxying()
  3.) Send the "host" message
  4.) Flush messages from the locked endpoint
  5.) Unlock the endpoint

so `PopulateEndpointDescriptor[...]()` is the wrong scope in which to take a
lock.

The goal was to eventually merge the endpoint-proxying logic inside
`PopulateEndpointDescriptor[...]()` with
`Node::SendMessagesAndRecursiveDependents()`, since that place was properly
locking dependent endpoints and flushing messages appropriately.

Another reason why we had to eventually merge the aforementioned two paths is:
in the case where a message-with-endpoints was being sent to a local endpoint
with a remote peer (not a local peer in the proxying state), the endpoint
population code was properly setting the dependent endpoint into the proxying
state, but no code was actually flushing already-queued messages from the
endpoint to the remote node, after its host message was sent, like
`Node::SendMessagesAndRecursiveDependents()` was doing. Therefore, when sending
an endpoing-with-queued-messages to a local endpoint with a remote peer, the
queued messages would never get delivered to the remote node.

This commit adds a test for the above scenario
(`MageTest.SendInvitationAndReceiveQueuedEndpointsFromAcceptor`) and fixes it by
calling `Node::SendMessagesAndRecursiveDependents()` in the peer-is-remote path
in `Node::SendMessage()`, thus allowing us to remove the endpoint-proxying logic
in `PopulateEndpointDescriptor[...]()`, which is a huge simplification!