Lock Core::handle_table_ for exclusive access
An occasional race condition could be found in
`MageTest.ChildPassSendInvitationPipeBackToParent`, where one thread was trying
to `find()` an entry *that should exist* in the handle table while another
thread is writing a new entry to the handle table, which was breaking the other
thread's `find()`.

This commit introduces a lock to `Core`, for exclusive access to the
thread-unsafe handle table.