Fix race condition in `Node::AcceptInvitation()`
AcceptInvitation() was starting the new `Channel` before inserting it into the
map. But when a channel is started, the IO thread can immediately read messages
and touch `Node` state: for example, `Node::OnReceivedInvitation()` can run,
which expects the channel to exist in `Node::node_channel_map_`. Therefore,
`Node::AcceptInvitation()` must first insert the channel into the map, and then
Start() it. This fixes the race condition without requiring any locking.