Message::GetEndpointDescriptors() returns pointers
This is necessary for solving the bug where a handle gets passed back
and forth multiple times between two processes, and gets create with the
same name in the same process, crashing. The fix for this will involve
re-generating a cross-node-endpoint-name for user messages that get
forwarded, and that is only possible if we can modify endpoint
descriptors that live in an existing message's buffer, which is what
this commit enables.