Reorder operations in TLFIO::WatchSocket()
This commit reorders the operations in TaskLoopForIO::WatchSocket() from being:

  1.) Register the socket with the kernel
  2.) Add the socket reader to the socket reader list

... to the exact opposite of this. If `WatchSocket()` was called on a thread
different from the IO thread, then the socket might become readable from the IO
thread's `Run()` method before the socket is registered in the list, causing the
IO thread's `Run()` crash when it tries to look up the associated socket reader
from the internal list.

In reality, we should probably just CHECK that WatchSocket() is called on the IO
thread instead of any other thread, but for now this will do.